IGAPP-669: Move assets
