+++ native/src/components/NewsDetail.tsx
@@ -4,7 +4,7 @@ import { ScrollView, useWindowDimensions, View } from 'react-native'
 import { LocalNewsModel, replaceLinks, TunewsModel } from 'api-client'
 import { ThemeType } from 'build-configs'
 import { contentAlignment } from '../constants/contentDirection'
-import headerImage from '../routes/assets/tu-news-header-details-icon.svg'
+import headerImage from '../assets/tu-news-header-details-icon.svg'
 import styled from 'styled-components/native'
 import Html from 'react-native-render-html'
 import TimeStamp from './TimeStamp'
