+++ native/src/components/NewsHeader.tsx
@@ -3,8 +3,8 @@ import { CityModel } from 'api-client'
 import styled from 'styled-components/native'
 import { withTranslation } from 'react-i18next'
 import withTheme from '../hocs/withTheme'
-import activeInternational from '../routes/assets/tu-news-active.svg'
-import inactiveInternational from '../routes/assets/tu-news-inactive.svg'
+import activeInternational from '../assets/tu-news-active.svg'
+import inactiveInternational from '../assets/tu-news-inactive.svg'
 import { TFunction } from 'i18next'
 import { LOCAL_NEWS_TYPE, TU_NEWS_TYPE, NewsType } from 'api-client/src/routes'
 import { ThemeType } from 'build-configs'
