+++ native/src/routes/Dashboard.tsx
@@ -5,11 +5,11 @@ import CategoriesRouteStateView from '../models/CategoriesRouteStateView'
 import { LanguageResourceCacheStateType } from '../redux/StateType'
 import NavigationTiles from '../components/NavigationTiles'
 import TileModel from '../models/TileModel'
-import eventsIcon from './assets/events.svg'
-import offersIcon from './assets/offers.svg'
-import poisIcon from './assets/pois.svg'
-import newsIcon from './assets/news.svg'
-import localInformationIcon from './assets/local_information.svg'
+import eventsIcon from '../assets/events.svg'
+import offersIcon from '../assets/offers.svg'
+import poisIcon from '../assets/pois.svg'
+import newsIcon from '../assets/news.svg'
+import localInformationIcon from '../assets/local_information.svg'
 import { TFunction } from 'react-i18next'
 import buildConfig from '../constants/buildConfig'
 import SpaceBetween from '../components/SpaceBetween'
